refactor(element): move tokens/lifecycle.js -> tokens.js
