+++ packages/element/src/index.js
@@ -3,5 +3,5 @@ export {createComputingToken, computer, observer} from './computer';
 export {default as element} from './element';
 export {default as internal} from './internal';
 export {default as property} from './property';
-export * from './tokens/lifecycle';
+export * from './tokens';
 export * from './query';
